+++ app.js
@@ -1,14 +1,12 @@
 const dotenv = require('dotenv');
 const logger = require('tracer').colorConsole();
-if(dotenv.config().error) {
+if (dotenv.config().error) {
     logger.error(`Failed to load dotenv`)
 
 }
 const express = require('express')
 const config = require('./config.json');
 const app = express()
-var bodyParser = require('body-parser')
-var jsonParser = bodyParser.json()
 const port = process.env.API_PORT || config.apiport;
 
 app.listen(port, () => {
